refactor(web): extract fake OpenCode e2e client
Move the smoke-only OpenCode adapter out of the production client factory so the runtime seam stays in one place without mixing in the fake implementation details.
